Changing main navigation to tab layout, adding SettingsFragment.kt

+++ app/src/main/java/de/zoghaib/schwimmbadguide/data/PoolInformations.kt
@@ -34,5 +34,5 @@ data class PoolInformations (
     val so1 : String,
     val so2 : String,
     val prices : String,
-    var distance : Int
+    var distance : Float
 )
